Updated typing and removed transform model and added it to controller
